renames director_v2_core_computations and moves DirectorV2API as ComputationsApi
