+++ services/web/server/src/simcore_service_webserver/director_v2_core_base.py
@@ -14,8 +14,6 @@ from typing import Any, Optional, Union
 
 import aiohttp
 from aiohttp import ClientTimeout, web
-from models_library.projects import ProjectID
-from models_library.users import UserID
 from tenacity._asyncio import AsyncRetrying
 from tenacity.before_sleep import before_sleep_log
 from tenacity.stop import stop_after_attempt
@@ -23,15 +21,10 @@ from tenacity.wait import wait_random
 from yarl import URL
 
 from .director_v2_exceptions import DirectorServiceError
-from .director_v2_settings import (
-    DirectorV2Settings,
-    get_client_session,
-    get_plugin_settings,
-)
+from .director_v2_settings import get_client_session
 
 log = logging.getLogger(__name__)
 
-_APP_DIRECTOR_V2_CLIENT_KEY = f"{__name__}.DirectorV2ApiClient"
 
 SERVICE_HEALTH_CHECK_TIMEOUT = ClientTimeout(total=2, connect=1)  # type:ignore
 
@@ -59,12 +52,13 @@ async def request_director_v2(
     ] = None,
     **kwargs,
 ) -> DataBody:
-    if not on_error:
-        on_error = {}
+    session = get_client_session(app)
+    on_error = on_error or {}
+
     try:
         async for attempt in AsyncRetrying(**DEFAULT_RETRY_POLICY):
             with attempt:
-                session = get_client_session(app)
+
                 async with session.request(
                     method, url, headers=headers, json=data, **kwargs
                 ) as response:
@@ -105,49 +99,3 @@ async def request_director_v2(
         reason="Unexpected client error",
         url=url,
     )
-
-
-class DirectorV2ApiClient:
-    def __init__(self, app: web.Application) -> None:
-        self._app = app
-        self._settings: DirectorV2Settings = get_plugin_settings(app)
-
-    async def get(self, project_id: ProjectID, user_id: UserID) -> dict[str, Any]:
-        computation_task_out = await request_director_v2(
-            self._app,
-            "GET",
-            (self._settings.base_url / "computations" / f"{project_id}").with_query(
-                user_id=int(user_id)
-            ),
-            expected_status=web.HTTPOk,
-        )
-        assert isinstance(computation_task_out, dict)  # nosec
-        return computation_task_out
-
-    async def start(self, project_id: ProjectID, user_id: UserID, **options) -> str:
-        computation_task_out = await request_director_v2(
-            self._app,
-            "POST",
-            self._settings.base_url / "computations",
-            expected_status=web.HTTPCreated,
-            data={"user_id": user_id, "project_id": project_id, **options},
-        )
-        assert isinstance(computation_task_out, dict)  # nosec
-        return computation_task_out["id"]
-
-    async def stop(self, project_id: ProjectID, user_id: UserID):
-        await request_director_v2(
-            self._app,
-            "POST",
-            self._settings.base_url / "computations" / f"{project_id}:stop",
-            expected_status=web.HTTPAccepted,
-            data={"user_id": user_id},
-        )
-
-
-def get_client(app: web.Application) -> Optional[DirectorV2ApiClient]:
-    return app.get(_APP_DIRECTOR_V2_CLIENT_KEY)
-
-
-def set_client(app: web.Application, obj: DirectorV2ApiClient):
-    app[_APP_DIRECTOR_V2_CLIENT_KEY] = obj
